// generated/src/app/api-models/check-clear-void-setup.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CheckClearVoidSetup {
  jobId: string;
  seqCvsupId: number;
  requestUser: string;
  requestDate: Date;
  action: string;
  daemonRequest: string;
  status: string;
  comments: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}